% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.character.GridPattern}
\alias{as.character.GridPattern}
\title{Character representation of a grid pattern object}
\usage{
\method{as.character}{GridPattern}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{arguments which may be passed to other functions}
}
\description{
Character representation of a grid pattern object
}
