% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CheckedData-class}
\alias{CheckedData-class}
\title{An S4 class to represent checked data.}
\description{
An S4 class to represent checked data.
}
\section{Slots}{

\describe{
\item{\code{inputdata}}{input data for CheckData function}
}}

