% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\docType{data}
\name{.mid}
\alias{.mid}
\title{Shortcut value for `unit(0.5, 'npc')`}
\format{
An object of class \code{simpleUnit} (inherits from \code{unit}, \code{unit_v2}) of length 1.
}
\usage{
.mid
}
\description{
Shortcut value for `unit(0.5, 'npc')`
}
\keyword{datasets}
