% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dotmm}
\alias{dotmm}
\alias{.mm}
\alias{.npc}
\alias{.inch}
\alias{.cm}
\alias{.points}
\alias{.snpc}
\title{dotmm}
\usage{
.mm(x)

.npc(x)

.inch(x)

.cm(x)

.points(x)

.snpc(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Specialized functions for unit creation
}
