% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneRegroup.R
\name{geneRegroup}
\alias{geneRegroup}
\title{geneRegroup function}
\usage{
geneRegroup(plist = TCGA$pathList)
}
\arguments{
\item{plist}{user provided gene lists organized in pathways}
}
\description{
redefining the gene set membership of common genes. First, the gene remains to be as a member if it is a core member of the pathway. Second, if the gene is not a core member of any among the pathways (i.e., it maps to more than one pathways), then this gene is re-assigned to one of the gene sets based on the k-medoids algorithm minimizing the binary distance between genes within cluster distance.
}
\examples{
data(TCGA)
geneRegroup.result=geneRegroup(plist=TCGA$pathList )
prefilter.results=prefilter( data=TCGA$geneexpr, time=TCGA$t, status=TCGA$d,
                             plist=geneRegroup.result@gset )
}
