% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpar.R
\name{gp}
\alias{gp}
\title{Simple wrapper for \code{grid::gpar()} which exposes the valid arguments}
\usage{
gp(
  col,
  fill,
  alpha,
  lty,
  lwd,
  lex,
  lineend,
  linejoin,
  linemitre,
  fontsize,
  cex,
  fontfamily,
  fontface,
  lineheight,
  ...
)
}
\arguments{
\item{col}{Colour for lines and borders.}

\item{fill}{Colour for filling rectangles, polygons, ...}

\item{alpha}{Alpha channel for transparency}

\item{lty}{Line type}

\item{lwd}{Line width}

\item{lex}{Multiplier applied to line width}

\item{lineend}{Line end style ('round', 'butt', 'square')}

\item{linejoin}{Line join style ('round', 'mitre', 'bevel')}

\item{linemitre}{Line mitre limit (number greater than 1)}

\item{fontsize}{The size of text (in points)}

\item{cex}{Multiplier applied to fontsize}

\item{fontfamily}{The font family}

\item{fontface}{The font face ('bold', 'italic', ...)}

\item{lineheight}{The height of a line as a multiple of the size of text}

\item{...}{other arguments ignored}
}
\description{
The \code{grid::gpar()} does not have any named arguments, and when used
in a modern IDE, it is difficult to quicklyl know which graphical parameters
are available.
}
\details{
This function (\code{ingrid::gp()}) is a thin wrapper around \code{grid::gpar()}
which exposes all the parameters as named arguments to make autocomplete
a bit more useful.
}
\examples{
\dontrun{
gp(fill = 'red')
}

}
