% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpar.R
\name{gp_update}
\alias{gp_update}
\title{Update the values in a gpar object}
\usage{
gp_update(
  gpar_obj,
  col,
  fill,
  alpha,
  lty,
  lwd,
  lex,
  lineend,
  linejoin,
  linemitre,
  fontsize,
  cex,
  fontfamily,
  fontface,
  lineheight
)
}
\arguments{
\item{gpar_obj}{gpar object}

\item{col}{Colour for lines and borders.}

\item{fill}{Colour for filling rectangles, polygons, ...}

\item{alpha}{Alpha channel for transparency}

\item{lty}{Line type}

\item{lwd}{Line width}

\item{lex}{Multiplier applied to line width}

\item{lineend}{Line end style ('round', 'butt', 'square')}

\item{linejoin}{Line join style ('round', 'mitre', 'bevel')}

\item{linemitre}{Line mitre limit (number greater than 1)}

\item{fontsize}{The size of text (in points)}

\item{cex}{Multiplier applied to fontsize}

\item{fontfamily}{The font family}

\item{fontface}{The font face ('bold', 'italic', ...)}

\item{lineheight}{The height of a line as a multiple of the size of text}
}
\description{
Update the values in a gpar object
}
