% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-combine.R
\name{grob_combine}
\alias{grob_combine}
\alias{grob_intersect}
\alias{grob_union}
\alias{grob_minus}
\alias{grob_xor}
\title{grob operations using \code{gridGeometry::polyclipGrob()}}
\usage{
grob_combine(
  op = c("intersection", "xor", "union", "minus"),
  x,
  y,
  gp = x$gp,
  name = NULL
)

grob_intersect(x, y, gp = x$gp, name = NULL)

grob_union(x, y, gp = x$gp, name = NULL)

grob_minus(x, y, gp = x$gp, name = NULL)

grob_xor(x, y, gp = x$gp, name = NULL)
}
\arguments{
\item{op}{operation. One of 'intersection', 'xor', 'union', 'minus'}

\item{x, y}{grobs, gtrees of polyclipgrob objects}

\item{gp}{gpar settings. default \code{x$gp}}

\item{name}{default NULL}
}
\description{
This wrapper sets the gp on the result to be the gp of the first argument.
}
