% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-utils.R
\name{grob_copy}
\alias{grob_copy}
\title{Copy a grob to a new named object}
\usage{
grob_copy(grob)
}
\arguments{
\item{grob}{grob}
}
\description{
TODO this should probably recursiely descend through gTrees and polyclipgrobs
and rename all child objects as well.
}
