% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-combine.R
\name{grob_invert}
\alias{grob_invert}
\title{Invert the geometry of a grob}
\usage{
grob_invert(x, name = NULL)
}
\arguments{
\item{x}{grob}

\item{name}{default: NULL}
}
\description{
Invert the geometry of a grob
}
