% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-transform.R
\name{grob_rotate}
\alias{grob_rotate}
\title{Rotate the (x,y) coordinates of a grob}
\usage{
grob_rotate(
  grob,
  angle,
  xc = unit(0.5, "npc"),
  yc = unit(0.5, "npc"),
  default.units = "npc"
)
}
\arguments{
\item{grob}{grob}

\item{angle}{rotation angle}

\item{xc, yc}{centre of rotation.}

\item{default.units}{'npc'}
}
\description{
Rotate the (x,y) coordinates of a grob
}
