% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-combine.R
\name{grob_stack}
\alias{grob_stack}
\title{Stack multiple grobs toegher into a grobTree}
\usage{
grob_stack(..., name = NULL)
}
\arguments{
\item{...}{grobs}

\item{name}{default: NULL}
}
\description{
Stack multiple grobs toegher into a grobTree
}
