% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-transform.R
\name{grob_translate}
\alias{grob_translate}
\title{Translate the (x,y) coordinates of a  grob}
\usage{
grob_translate(
  grob,
  x,
  y,
  default.units = getOption("ingrid.default.units", "mm")
)
}
\arguments{
\item{grob}{grob}

\item{x, y}{translation}

\item{default.units}{getOption("ingrid.default.units", 'mm')}
}
\description{
Translate the (x,y) coordinates of a  grob
}
