% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-combine.R
\name{igc_combine}
\alias{igc_combine}
\alias{igc_intersect}
\alias{igc_union}
\alias{igc_minus}
\alias{igc_xor}
\title{grob operations using \code{gridGeometry::polyclipGrob()}}
\usage{
igc_combine(
  x,
  y,
  op = c("intersection", "xor", "union", "minus"),
  gp = x$gp,
  name = NULL
)

igc_intersect(x, y, gp = x$gp, name = NULL)

igc_union(x, y, gp = x$gp, name = NULL)

igc_minus(x, y, gp = x$gp, name = NULL)

igc_xor(x, y, gp = x$gp, name = NULL)
}
\arguments{
\item{x, y}{grobs, gtrees of polyclipgrob objects}

\item{op}{operation. One of 'intersection', 'xor', 'union', 'minus'}

\item{gp}{gpar settings. default \code{x$gp}}

\item{name}{default NULL}
}
\description{
This wrapper sets the graphical parameters on the result to be those of
the \code{x} argument.
}
\seealso{
Other combinators: 
\code{\link{igc_invert}()},
\code{\link{igc_stack}()}
}
\concept{combinators}
