% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-combine.R
\name{igc_invert}
\alias{igc_invert}
\title{Invert the geometry of a grob}
\usage{
igc_invert(x, name = NULL)
}
\arguments{
\item{x}{grob}

\item{name}{default: NULL}
}
\description{
Invert the geometry of a grob
}
\seealso{
Other combinators: 
\code{\link{igc_combine}()},
\code{\link{igc_stack}()}
}
\concept{combinators}
