% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewport.R
\name{igc_mask}
\alias{igc_mask}
\title{Set a mask on an existing viewport or grob}
\usage{
igc_mask(x, mask)
}
\arguments{
\item{x}{viewport or grob with a viewport}

\item{mask}{One of "none" (or FALSE) or "inherit" (or TRUE) or a grob
(or a gTree). This specifies that the viewport should have no mask,
or it should inherit the mask of its parent, or it should have its
own mask, as described by the grob.}
}
\description{
Set a mask on an existing viewport or grob
}
