% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-combine.R
\name{igc_stack}
\alias{igc_stack}
\title{Stack multiple grobs toegher into a grobTree}
\usage{
igc_stack(..., name = NULL)
}
\arguments{
\item{...}{grobs}

\item{name}{default: NULL}
}
\description{
Stack multiple grobs toegher into a grobTree
}
\seealso{
Other combinators: 
\code{\link{igc_combine}()},
\code{\link{igc_invert}()}
}
\concept{combinators}
