% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{igl_row}
\alias{igl_row}
\alias{igl_col}
\title{Functions for laying out multiple grobs}
\usage{
igl_row(..., widths = NULL)

igl_col(..., heights = NULL)
}
\arguments{
\item{...}{grobs}

\item{widths, heights}{the relative widths or heights of the grobs for the
\code{row()} and and \code{col()} layouts respectively.  Default: NULL
means to allocate equal space to all grobs}
}
\value{
a grob object combining the given grobs
}
\description{
Functions for laying out multiple grobs
}
\seealso{
Other layout: 
\code{\link{igl_vp}()}
}
\concept{layout}
