% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{igl_vp}
\alias{igl_vp}
\title{Nested layout with an arbitrary \code{grid::viewport()}}
\usage{
igl_vp(
  ...,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  width = unit(1, "npc"),
  height = unit(1, "npc"),
  default.units = "npc",
  just = "centre",
  gp = gpar(),
  clip = "inherit",
  mask = "inherit",
  xscale = c(0, 1),
  yscale = c(0, 1),
  angle = 0,
  layout = NULL,
  layout.pos.row = NULL,
  layout.pos.col = NULL,
  name = NULL
)
}
\arguments{
\item{...}{grobs within this viewport}

\item{x, y, width, height, default.units, just, gp, clip, mask, xscale, yscale, angle, layout, layout.pos.row, layout.pos.col, name}{See \code{grid::viewport()} documentation.}
}
\description{
Nested layout with an arbitrary \code{grid::viewport()}
}
\seealso{
Other layout: 
\code{\link{igl_row}()}
}
\concept{layout}
