% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-transform.R
\name{igt_rotate}
\alias{igt_rotate}
\title{Rotate a grob or viewport}
\usage{
igt_rotate(obj, angle)
}
\arguments{
\item{obj}{grob or viewport}

\item{angle}{rotation angle in degrees}
}
\description{
Rotate a grob or viewport
}
\seealso{
Other transformations: 
\code{\link{igt_translate}()},
\code{\link{igt_update}()}
}
\concept{transformations}
