% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-transform.R
\name{igt_translate}
\alias{igt_translate}
\title{Translate a grob or viewport}
\usage{
igt_translate(
  obj,
  x = 0,
  y = 0,
  default.units = getOption("ingrid.default.units", "npc")
)
}
\arguments{
\item{obj}{grob or viewport}

\item{x, y}{translation}

\item{default.units}{'npc'}
}
\description{
Translate a grob or viewport
}
\seealso{
Other transformations: 
\code{\link{igt_rotate}()},
\code{\link{igt_update}()}
}
\concept{transformations}
