% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-transform.R
\name{igt_update}
\alias{igt_update}
\title{Update a grob/viewport}
\usage{
igt_update(obj, ...)
}
\arguments{
\item{obj}{grob or viewport}

\item{...}{named arguments e.g. \code{fill = 'red'}}
}
\description{
Note: No checking is currently done on the validity of these parameters.
      Use with caution!
}
\seealso{
Other transformations: 
\code{\link{igt_rotate}()},
\code{\link{igt_translate}()}
}
\concept{transformations}
