% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-utils.R
\name{is_vanilla_grob}
\alias{is_vanilla_grob}
\alias{is_polyclipgrob}
\alias{is_vanilla_gTree}
\alias{is_pattern}
\alias{is_grob}
\alias{is_viewport}
\title{Testing for grob types}
\usage{
is_vanilla_grob(x)

is_polyclipgrob(x)

is_vanilla_gTree(x)

is_pattern(x)

is_grob(x)

is_viewport(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Testing for grob types
}
