% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob.R
\name{path_grob}
\alias{path_grob}
\title{Create a grob}
\usage{
path_grob(
  x = c(0, 1, 1) * 20,
  y = c(0, 0, 1) * 20,
  id = NULL,
  id.lengths = NULL,
  pathId = NULL,
  pathId.lengths = NULL,
  rule = c("winding", "evenodd"),
  default.units = getOption("ingrid.default.units", "mm"),
  name = NULL,
  gp = NULL,
  vp = NULL,
  centred = TRUE,
  col,
  fill,
  alpha,
  lty,
  lwd,
  lex,
  lineend,
  linejoin,
  linemitre,
  fontsize,
  cex,
  fontfamily,
  fontface,
  lineheight,
  mask,
  clip,
  layout,
  layout.pos.row,
  layout.pos.col
)
}
\arguments{
\item{x}{x,y location}

\item{y}{x,y location}

\item{id}{A numeric vector used to separate locations in x and y into
sub-objects. All locations with the same id belong to the same
sub-object.}

\item{id.lengths}{A numeric vector used to separate locations in x and y
into sub-objects. Specifies consecutive blocks of locations which
 make up separate sub-objects.}

\item{pathId}{A numeric vector used to separate locations in x and y
into distinct objects. All locations with the same pathId belong
to the same object.}

\item{pathId.lengths}{A numeric vector used to separate locations in
x and y into objects. Specifies consecutive blocks of locations
which make up separate objects.}

\item{rule}{A character value specifying the fill rule: either
"winding" or "evenodd".}

\item{default.units}{\code{getOption("ingrid.default.units", 'mm')}}

\item{name}{grob name. default: NULL}

\item{gp}{graphical parameter object created by \code{grid::gpar()} or
\code{ingrid::gp()}.  If NULL (default), then a \code{gpar} object
is created from the relevant arguments to this function i.e. \code{fill},
\code{col}, etc}

\item{vp}{viewport object created by \code{grid::viewport()} or
\code{ingrid::vp()}.  If NULL (default), then a \code{viewport} object
is created from a limited subset of arguments to this function i.e. \code{mask},
\code{clip}, etc}

\item{centred}{default: TRUE to shift given x,y coordinates by
c(0.5npc, 0.5npc)}

\item{col}{See documentation for \code{gp()}}

\item{fill}{See documentation for \code{gp()}}

\item{alpha}{See documentation for \code{gp()}}

\item{lty}{See documentation for \code{gp()}}

\item{lwd}{See documentation for \code{gp()}}

\item{lex}{See documentation for \code{gp()}}

\item{lineend}{See documentation for \code{gp()}}

\item{linejoin}{See documentation for \code{gp()}}

\item{linemitre}{See documentation for \code{gp()}}

\item{fontsize}{See documentation for \code{gp()}}

\item{cex}{See documentation for \code{gp()}}

\item{fontfamily}{See documentation for \code{gp()}}

\item{fontface}{See documentation for \code{gp()}}

\item{lineheight}{See documentation for \code{gp()}}

\item{mask}{See documentation for
\code{grid::viewport()}}

\item{clip}{See documentation for
\code{grid::viewport()}}

\item{layout}{See documentation for
\code{grid::viewport()}}

\item{layout.pos.row}{See documentation for
\code{grid::viewport()}}

\item{layout.pos.col}{See documentation for
\code{grid::viewport()}}
}
\description{
Create a grob
}
