% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern_create}
\alias{pattern_create}
\title{Alternate call for \code{grid::pattern()} with different defaults}
\usage{
pattern_create(
  grob,
  x = 0,
  y = 0,
  width = 20,
  height = 20,
  default.units = getOption("ingrid.default.units", "mm"),
  just = "centre",
  hjust = NULL,
  vjust = NULL,
  extend = c("repeat", "pad", "reflect", "none"),
  gp = gpar(fill = "transparent"),
  centred = TRUE
)
}
\arguments{
\item{grob}{grob}

\item{x, y}{default 0,0}

\item{width, height}{default: 20,20}

\item{default.units}{getOption("ingrid.default.units", 'mm')}

\item{just}{justification. default: centre}

\item{hjust, vjust}{other justification parameters}

\item{extend}{default: 'repeat'}

\item{gp}{graphical parameters}

\item{centred}{default: TRUE. (x, y) will be offset from (0.5npc, 0.5npc)}
}
\description{
Alternate call for \code{grid::pattern()} with different defaults
}
