% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern_view}
\alias{pattern_view}
\title{view a pattern}
\usage{
pattern_view(pat, height = .mm(50), width = .mm(50 * 1.618), clear = TRUE, ...)
}
\arguments{
\item{pat}{pattern object}

\item{width, height}{dimenaions of rectangle holding pattern}

\item{clear}{should the page be clared before drawing pattern? default: TRUE}

\item{...}{other arguments passed to \code{rect_grob}}
}
\description{
view a pattern
}
