% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.GridLinearGradient}
\alias{print.GridLinearGradient}
\alias{print.GridRadialGradient}
\alias{print.GridTilingPattern}
\alias{print.GridMask}
\title{Print a pattern}
\usage{
\method{print}{GridLinearGradient}(x, depth = 1L, prefix = NULL, ...)

\method{print}{GridRadialGradient}(x, depth = 1L, prefix = NULL, ...)

\method{print}{GridTilingPattern}(x, depth = 1L, prefix = NULL, ...)

\method{print}{GridMask}(x, depth = 1L, prefix = NULL, ...)
}
\arguments{
\item{x}{object to print}

\item{depth}{indentation depth. Default: 1L}

\item{prefix}{Prefix for first line of output. Default: NULL}

\item{...}{arguments which may be passed to other functions}
}
\description{
Print a pattern
}
