% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.grob}
\alias{print.grob}
\title{print a grob a bit more verbose}
\usage{
\method{print}{grob}(x, depth = 1L, include_gp = TRUE, include_vp = TRUE, prefix = NULL, ...)
}
\arguments{
\item{x}{object to print}

\item{depth}{indentation depth. Default: 1L}

\item{include_gp, include_vp}{print the graphics parameters and viewport?
default: TRUE for both}

\item{prefix}{Prefix for first line of output. Default: NULL}

\item{...}{arguments which may be passed to other functions}
}
\description{
print a grob a bit more verbose
}
