% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gTree}
\alias{print.gTree}
\title{print a grob a bit more verbose}
\usage{
\method{print}{gTree}(x, depth = 1L, prefix = NULL, ...)
}
\arguments{
\item{x}{object to print}

\item{depth}{indentation depth. Default: 1L}

\item{prefix}{Prefix for first line of output. Default: NULL}

\item{...}{arguments which may be passed to other functions}
}
\description{
print a grob a bit more verbose
}
