% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{register_verbose_printing}
\alias{register_verbose_printing}
\title{Register the verbose printing methods for grobs, polyclipgrobs, gTrres, gpars, and viewports}
\usage{
register_verbose_printing()
}
\description{
These will all override the standard print methods in the \code{grid} package.
}
