% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RegroupGene-class}
\alias{RegroupGene-class}
\title{An S4 class to represent regrouped gene set.}
\description{
An S4 class to represent regrouped gene set.
}
\section{Slots}{

\describe{
\item{\code{gset}}{list of regrouped gene set}

\item{\code{inputdata}}{input data for geneRegroup function}
}}

