% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGene.R
\docType{methods}
\name{selectGene,Prefiltered-method}
\alias{selectGene}
\alias{selectGene,Prefiltered-method}
\title{Gene selection using SPLS model}
\usage{
\S4method{selectGene}{Prefiltered}(object, K = 5, etas = seq(0.1, 0.9, 0.1),
  fold = 5, se1 = TRUE, method = "plik", par = FALSE, foldid = NULL,
  seed = 123)
}
\arguments{
\item{object}{output list of prefilter step}

\item{K}{the maximum number of hidden features in spls.}

\item{etas}{Thresholding parameter. eta should be between 0 and 1.}

\item{fold}{The number of folds to use to perform the cross-validation process.}

\item{se1}{one standard error criteria, default = TRUE}

\item{method}{if auc is used, this is the maximum auc criteria, if partial likelihood(plik) is used, this is the min deviance criteria. Default = TRUE}

\item{par}{parallel option for running the code, default = FALSE}

\item{foldid}{the number of folds to use to perform the cross-validation process.}

\item{seed}{random seed that was set,  default = 123.}
}
\description{
Sparse Partial Least Square (SPLS) step for gene selection and dimension reduction.By applying SPLS to each pathway, we achieve the goal of gene selection and dimension reduction at the same time.
}
\examples{
data(TCGA)
geneRegroup.result=geneRegroup(plist=TCGA$pathList )
prefilter.results=prefilter( data=TCGA$geneexpr, time=TCGA$t, status=TCGA$d,
                             plist=geneRegroup.result@gset )
gene.results=selectGene( object=prefilter.results, fold=5, K=5, etas=c(0.1,0.9),seed=123)
}

