% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DataCheck.R
\docType{methods}
\name{show,CheckedData-method}
\alias{show,CheckedData-method}
\title{checkData summary}
\usage{
\S4method{show}{CheckedData}(object)
}
\arguments{
\item{object}{output of prefilter function}
}
\description{
Summary of checkData results.
}
