% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_RegroupGene.R
\docType{methods}
\name{show,RegroupGene-method}
\alias{show,RegroupGene-method}
\alias{show,geneRegroup-method}
\title{Gene Regrouping summary}
\usage{
\S4method{show}{RegroupGene}(object)
}
\arguments{
\item{object}{output of geneRegroup function}
}
\description{
Summary of gene Regrouping results.
}
