% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_full.R
\docType{data}
\name{TCGA_full}
\alias{TCGA_full}
\title{Full TCGA data}
\format{A \code{list} object with four elements, \code{geneexpr} (z scores for gene expression data), \code{t} (survival time), \code{d} (survival status indicator), \code{pathList} (pathway information).}
\usage{
TCGA_full
}
\description{
Full TCGA data
}
\examples{
data(TCGA_full)
}
\references{
TCGA_full
}
\keyword{datasets}
