% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewport.R
\name{vp_setclip}
\alias{vp_setclip}
\title{Set a clip on an existing viewport}
\usage{
vp_setclip(vp, clip)
}
\arguments{
\item{vp}{viewport}

\item{clip}{One of "on", "inherit", or "off", indicating whether to clip
to the extent of this viewport, inherit the clipping region from the
parent viewport, or turn clipping off altogether. For back-compatibility,
 a logical value of TRUE corresponds to "on" and FALSE corresponds to
 "inherit". May also be a grob (or a gTree) that describes a clipping path.}
}
\description{
Set a clip on an existing viewport
}
