% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewport.R
\name{vp_setmask}
\alias{vp_setmask}
\title{Set a mask on an existing viewport}
\usage{
vp_setmask(vp, mask)
}
\arguments{
\item{vp}{viewport}

\item{mask}{One of "none" (or FALSE) or "inherit" (or TRUE) or a grob
(or a gTree). This specifies that the viewport should have no mask,
or it should inherit the mask of its parent, or it should have its
own mask, as described by the grob.}
}
\description{
Set a mask on an existing viewport
}
