% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewport.R
\name{vp_translate}
\alias{vp_translate}
\title{Translate a viewport}
\usage{
vp_translate(vp, x, y, default.units = "npc")
}
\arguments{
\item{vp}{viewport}

\item{x, y}{translation}

\item{default.units}{'npc'}
}
\description{
Translate a viewport
}
