% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewport.R
\name{vpc}
\alias{vpc}
\title{Thing wrapper around \code{grid::viewport()}}
\usage{
vpc(
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  width = unit(1, "npc"),
  height = unit(1, "npc"),
  default.units = "npc",
  just = "centre",
  gp = gpar(),
  clip = "inherit",
  mask = "inherit",
  xscale = c(0, 1),
  yscale = c(0, 1),
  angle = 0,
  layout = NULL,
  layout.pos.row = NULL,
  layout.pos.col = NULL,
  name = NULL,
  ...
)
}
\arguments{
\item{x, y, width, height, default.units, just, gp, clip, mask, xscale, yscale, angle, layout, layout.pos.row, layout.pos.col, name}{See \code{grid::viewport()} documentation.}

\item{...}{other arguments ignored}
}
\description{
Thing wrapper around \code{grid::viewport()}
}
